Small change made for array length and exit choice
In the menu class, there were some situations that can lead to errors. For instance, when the array length is inputted as a string or a non-positive integer an error occurs in the main. In addition, we added a small message for the exit choice for a better user experience.